+++ packages/match/configure/src/answer-config-block.jsx
@@ -8,8 +8,9 @@ import { withStyles } from '@material-ui/core/styles';
 import { InputCheckbox } from '@pie-lib/config-ui';
 import AddRow from './add-row';
 import Row from './row';
-import { swap } from '@pie-lib/drag';
+import { swap, withDragContext } from '@pie-lib/drag';
 import debug from 'debug';
+import lodash from 'lodash';
 
 const log = debug('pie-elements:match:configure');
 
@@ -49,16 +50,6 @@ const styles = theme => ({
     borderTop: '2px solid lightgray',
     width: '100%'
   },
-  checkboxContainer: {
-    display: 'flex',
-    justifyContent: 'space-between'
-  },
-  optionsCheckbox: {
-    flex: 1,
-    display: 'flex',
-    flexWrap: 'wrap',
-    justifyContent: 'space-between'
-  },
   headerInput: {
     textAlign: 'center',
     maxWidth: '100px'
@@ -94,13 +85,18 @@ class AnswerConfigBlock extends React.Component {
     onChange(newModel);
   };
 
-  onChange = (name, isBoolean) => event => {
+  onChange = (name, isBoolean) => ({ target }) => {
     const { model, onChange } = this.props;
-    const newModel = { ...model };
+    let value;
 
-    newModel[name] = isBoolean ? event.target.checked : event.target.value;
+    if (isBoolean) {
+      value = target.checked;
+    } else {
+      value = target.value;
+    }
 
-    onChange(newModel, name);
+    lodash.set(model, name, value);
+    onChange(model, name);
   };
 
   onHeaderChange = headerIndex => event => {
@@ -114,6 +110,13 @@ class AnswerConfigBlock extends React.Component {
 
   render() {
     const { classes, model, onAddRow, imageSupport } = this.props;
+    const {
+      configure: {
+        headers,
+        lockChoiceOrder,
+        partialScoring
+      }
+    } = model;
 
     return (
       <div className={classes.container}>
@@ -122,7 +125,7 @@ class AnswerConfigBlock extends React.Component {
         </Typography>
         <div className={classes.rowTable}>
           <div className={classes.rowContainer}>
-            {model.headers.map((header, idx) => (
+            {headers.settings && model.headers.map((header, idx) => (
               <div key={idx} className={cx(classes.rowItem, { [classes.questionText]: idx === 0 })}>
                 <Input
                   type="text"
@@ -151,27 +154,14 @@ class AnswerConfigBlock extends React.Component {
               onChange={this.props.onChange}
               onMoveRow={this.moveRow}
               imageSupport={imageSupport}
+              enableImages={model.enableImages}
             />
           ))}
           <AddRow onAddClick={onAddRow} />
-          <div className={classes.checkboxContainer}>
-            <div className={classes.optionsCheckbox}>
-              <InputCheckbox
-                label="Shuffle Choices"
-                checked={!!model.shuffled}
-                onChange={this.onChange('shuffled', true)}/>
-            </div>
-            <div className={classes.optionsCheckbox}>
-              <InputCheckbox
-                label="Allow Partial Scoring"
-                checked={!!model.allowPartialScoring}
-                onChange={this.onChange('allowPartialScoring', true)}/>
-            </div>
-          </div>
         </div>
       </div>
     );
   }
 }
 
-export default withStyles(styles)(AnswerConfigBlock);
+export default withDragContext(withStyles(styles)(AnswerConfigBlock));