+++ packages/pie-models/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@1.0.0...@pie-element/pie-models@1.0.1) (2019-04-10)
+
+**Note:** Version bump only for package @pie-element/pie-models
+
+
+
+
+
 # [1.0.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@0.5.4...@pie-element/pie-models@1.0.0) (2019-04-09)