+++ packages/placement-ordering/controller/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering-controller@3.1.0...@pie-element/placement-ordering-controller@3.1.1) (2019-04-11)
+
+**Note:** Version bump only for package @pie-element/placement-ordering-controller
+
+
+
+
+
+# [3.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering-controller@3.0.0...@pie-element/placement-ordering-controller@3.1.0) (2019-04-10)
+
+
+### Features
+
+* scoring change [ch440] ([42507f5](https://github.com/pie-framework/pie-elements/commit/42507f5))
+
+
+
+
+
 # [3.0.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering-controller@2.1.4...@pie-element/placement-ordering-controller@3.0.0) (2019-04-09)